chore: fix types error
